Add matrices to describe transformation to AC masks
These matrices correspond to transformations from pixel to plate for each
AC on the plate. They are created from the expected position of the AC
mask at the focal plane.

This assumes that the central AC dot is in the centre of the image, that
the AC mask isn’t tilted with respect to the yz plane, and that a pixel
corresponds to about 0.049255 mm (taken this from previous measurements).

These are NOT the final transformation matrices which will be created
using real measurements of the populated plate.